wallet send token API and wallet app seperation
