+++ backend/redlight_finance_37205/settings.py
@@ -73,6 +73,7 @@ INSTALLED_APPS = [
 LOCAL_APPS = [
     'home',
     'users.apps.UsersConfig',
+    'wallet'
 ]
 THIRD_PARTY_APPS = [
     'rest_framework',
