+++ backend/wallet/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class WalletConfig(AppConfig):
+    name = 'wallet'
